updated description and line 24
I had sent an email asking if the one-to-one mapping should follow the definition of the one-to-one function, meaning the output (s2) can be longer than the input(s1) and still be one-to-one if every character in s1 maps one of s2. Since I had not received an email in response before the due date, I submitted the assessment using the definition of one-to-one. Today I received the email responding my inquiry, they told me I needed both s1 and s2 to be equal, therefore not following the definition of one-to-one. The only modification needed for my code to comply to that requirement was change line 24 from an > sign to an == sign which is the only modification I made aside from the description to adjust to this slight change.